# Conflicts:
#	src/main/java/edu/wpi/veganvampires/controllers/Controller.java
#	src/main/java/edu/wpi/veganvampires/controllers/LaundryRequestController.java
#	src/main/java/edu/wpi/veganvampires/dao/LaundryRequestDao.java
#	src/main/java/edu/wpi/veganvampires/interfaces/LaundryRequestImpl.java
#	src/main/java/edu/wpi/veganvampires/objects/LaundryRequest.java
#	src/main/java/edu/wpi/veganvampires/objects/ServiceRequest.java
#	src/main/resources/FXML/LocationDB.fxml

+++ src/main/java/edu/wpi/veganvampires/objects/Location.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 public class Location {
   protected final int ID = 0; // location class that stores info about each location
   private String nodeID;
-  private int xCoord;
-  private int yCoord;
+  private double xCoord;
+  private double yCoord;
   private String floor;
   private String building;
   private String nodeType;
@@ -31,8 +31,8 @@ public class Location {
 
   public Location(
       String nodeID,
-      int xCoord,
-      int yCoord,
+      double xCoord,
+      double yCoord,
       String floor,
       String building,
       String nodeType,
